requirements.txt
django
djangorestframework
Pillow
django-cors-headers
python-dateutil
djangorestframework-simplejwt
dj-rest-auth